Update GamePieceClassifer.py
fixed an issue where it was not detecting the class that I wanted to train it on (pictures of consoles). Also added the read_image import from the torchvision.io library so that it can read images directly from the folder without having to look for sub-folders. Also when entering the directory, make sure to add two "/" at the first occurrence, because "/" acts as a 'stop' in python essentially, so you need two of them to cancel it out. Python is stupid like that.