Update FLAGSHIP.py
added other tof thresholds, 

NEED TO DO: 
Add corrective code for when we are descending the ramp in case we do not go down straight. 

Interpret string for color sensor and for the time of flight sensors for us to be able to start and for us to correct ourselves.